@@ -32,16 +32,16 @@
         }
     ],
     "require": {
-        "php": ">=7.0",
-        "illuminate/events": "~5.5 || ~6.0 || ~7.0 || ~8.0",
-        "illuminate/notifications": "~5.5 || ~6.0 || ~7.0 || ~8.0",
-        "illuminate/queue": "~5.5 || ~6.0 || ~7.0 || ~8.0",
-        "illuminate/support": "~5.5 || ~6.0 || ~7.0 || ~8.0",
+        "php": ">=7.4",
+        "illuminate/events": "~7.0 || ~8.0 || ~9.0",
+        "illuminate/notifications": "~7.0 || ~8.0 || ~9.0",
+        "illuminate/queue": "~7.0 || ~8.0 || ~9.0",
+        "illuminate/support": "~7.0 || ~8.0 || ~9.0",
         "pusher/pusher-push-notifications": "^1.1"
     },
     "require-dev": {
         "mockery/mockery": "^1.3",
-        "phpunit/phpunit": "^8.5"
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
